fix: moved hydrate to bin
