+++ package.json
@@ -13,8 +13,9 @@
 	},
 	"type": "module",
 	"main": "./lib/index.js",
-	"bin": "./lib/hydrate/index.js",
+	"bin": "./bin/hydrate.js",
 	"files": [
+		"bin/",
 		"lib/",
 		"package.json",
 		"LICENSE.md",
